Add RayCluster support for DGX Cloud Lepton (#389)
* Add Ray support for DGX Cloud Lepton

Add support for launching a RayCluster on DGX Cloud Lepton and submitting
RayJobs on the clusters using the lepton SDK. This uses the new RayCluster
feature on DGX Cloud Lepton to dynamically spawn clusters up and down via
the Python SDK and jobs can be submitted to deployed clusters directly.

Signed-Off-By: Robert Clark <roclark@nvidia.com>

* making name unique and add resource shape for lepton raycluster

Signed-off-by: Zoey Zhang <zozhang@nvidia.com>

* adding head node reference in RayCluster, support defining secrets in RayCluster and linting

Signed-off-by: Zoey Zhang <zozhang@nvidia.com>

* Remove Slurm packager comments from RayCluster

Removed the placeholder Slurm packager handling comments from the Lepton
RayCluster code. For now, the "workdir" parameter should be used for
transferring local data to the remote Ray cluster.

Signed-Off-By: Robert Clark <roclark@nvidia.com>

* Fix RayCluster head resource shape

Fix issue to ensure the proper head node resource shape is used if it
isn't explicitly given by the user.

Signed-Off-By: Robert Clark <roclark@nvidia.com>

* Update LeptonRay comments

Updated the comments in the LeptonRayCluster and LeptonRayJob classes to
accurately reflect the code.

Signed-Off-By: Robert Clark <roclark@nvidia.com>

* Fix RayJob logs streaming connection dropping

The RayJob logs stream would sometimes timeout and reset, causing a very
long output of logs in the terminal as it continually resets.

Signed-Off-By: Robert Clark <roclark@nvidia.com>

* Make RayCluster head resource shape optional

The head node resource shape for a LeptonRayCluster should be optional. If
it isn't specified by the user, it should default to the same shape used
for the worker nodes.

Signed-Off-By: Robert Clark <roclark@nvidia.com>

* Add doc for DGXC Lepton RayClusters

Added an example to the Ray quick-start guide on how to use RayClusters
and RayJobs with NeMo-Run on DGX Cloud Lepton.

Signed-Off-By: Robert Clark <roclark@nvidia.com>

* Update license date

Signed-Off-By: Robert Clark <roclark@nvidia.com>

* Fix Ray guide typo

Signed-Off-By: Robert Clark <roclark@nvidia.com>

* Make cluster readiness timeout a variable

Allows users to specify how long to wait for a RayCluster to be created
on DGX Cloud Lepton.

Signed-Off-By: Robert Clark <roclark@nvidia.com>

* Remove implicit returns

Signed-Off-By: Robert Clark <roclark@nvidia.com>

* Remove unused local variable

Signed-Off-By: Robert Clark <roclark@nvidia.com>

* Fix linting errors

Signed-Off-By: Robert Clark <roclark@nvidia.com>

* Fix formatting errors

Signed-Off-By: Robert Clark <roclark@nvidia.com>

* Move LeptonExecutor parameters to definition

Move the RayCluster-specific settings to the LeptonExecutor class for a
more seamless interface for launching and interacting with RayClusters
on DGX Cloud Lepton.

Signed-Off-By: Robert Clark <roclark@nvidia.com>

* Updated leptonai package version

Need a newer version of the leptonai SDK to support RayClusters.

Signed-Off-By: Robert Clark <roclark@nvidia.com>

* Add Lepton RayCluster tests

Signed-Off-By: Robert Clark <roclark@nvidia.com>

---------

Signed-off-by: Robert Clark <roclark@nvidia.com>
Signed-off-by: Zoey Zhang <zozhang@nvidia.com>
Co-authored-by: Zoey Zhang <zozhang@nvidia.com>